requirements.txt
fastapi[all]
authlib
requests
jose
itsdangerous
uvicorn
httpx
jinja2
python-multipart
pydantic[email]
sqlalchemy
asyncpg
psycopg2-binary
alembic
asyncpg
psycopg2-binary
alembic
termcolor
